+++ src/onegov.gazette/onegov/gazette/views/notice.py
@@ -32,7 +32,7 @@ def view_notice(self, request):
 
     actions = []
 
-    if request.is_personal(self) and not request.is_private(self):
+    if request.is_personal(self):
         user_ids, group_ids = get_user_and_group(request)
         if (self.group_id in group_ids) or (self.user_id in user_ids):
             if self.state == 'drafted' or self.state == 'rejected':